console: swap out the goja console to nodejs (experiment)
