+++ .gitignore
@@ -14,6 +14,7 @@
 */**/*tx_database*
 */**/*dapps*
 build/_vendor/pkg
+*/**/node_modules
 
 #*
 .#*
