+++ console-legacy/bridge.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with the go-ethereum library. If not, see <http://www.gnu.org/licenses/>.
 
-package console
+package console_legacy
 
 import (
 	"encoding/json"
@@ -28,7 +28,7 @@ import (
 	"github.com/dop251/goja"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/karalabe/tinygeth/console/prompt"
+	"github.com/karalabe/tinygeth/console-legacy/prompt"
 	"github.com/karalabe/tinygeth/internal/jsre"
 )
