+++ console-legacy/testdata/preload.js
