+++ console/package.json
@@ -1,0 +1,9 @@
+{
+  "dependencies": {
+    "chalk": "^5.3.0",
+    "ethers": "^6.13.4",
+    "pretty-repl": "^4.0.1",
+    "viem": "^2.21.32",
+    "web3": "^4.14.0"
+  }
+}
